add function description and clean up code
